LayoutTests/http/tests/xmlhttprequest/inject-header-expected.txt

Test that setRequestHeader properly checks for line feeds in header values.

a%0AEvil%3A%20on -> SUCCESS, setRequestHeader() raised an exception Error: SyntaxError: DOM Exception 12

a%0DEvil%3A%20on -> SUCCESS, setRequestHeader() raised an exception Error: SyntaxError: DOM Exception 12

a%0D%0AEvil%3A%20on -> SUCCESS, setRequestHeader() raised an exception Error: SyntaxError: DOM Exception 12

a%0A%0DEvil%3A%20on -> SUCCESS, setRequestHeader() raised an exception Error: SyntaxError: DOM Exception 12
